requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
boto==2.49.0
boto3==1.9.100
botocore==1.12.100
Click==7.0
coverage==4.5.4
csvvalidator==1.2
Django==2.2.4
docutils==0.14
Flask==1.1.1
image==1.5.27
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.16.2
packaging==19.1
pandas==0.24.2
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
